Add accelerator support for packing and unpacking
Support all types of datatype, including vector, indexed and struct in
addition to contiguous. The packing is done by a kernel that is
currently launched for each operation.

Basically, the datatype description is transposed into a flatten iovec,
and this list of iovecs is i[evenly] distributed amoung all the SM used
for the kernel.

Add a tests to validate the packing/unpacking.

Signed-off-by: George Bosilca <gbosilca@nvidia.com>