docs: update list of deprecated/removed functions
This commit address issue #10099

Update documentation with the list of depreacted functions in MPI 2.2
and MPI-4.0. Add some examples and explanation on how to convert
code from the deprecated interfaces to the new ones.

Add a new section on ABI compatibility to Open MPI 4.1
Add text about Fortran ABI issues.

Signed-off-by: Edgar Gabriel <edgar.gabriel@amd.com>
Co-authored-by: Jeff Squyres <jeff@squyres.com>