docs: Better explain required versions of support libraries
Get rid of the table that listed the required support library versions
-- the table format was too limiting.  Instead, add a bunch more
verbiage about the versions that are included in the Open MPI
distribution tarball and the minimum required versions of each of
hwloc, libevent, PMIx, and PRRTE.  Pay special attention to the corner
cases of building PMIx (internal and external), with and without
PRRTE.

Also set the minimum required versions for PMIx and PRRTE in VERSIONS:

* Testing shows that Open MPI v5.0.x requires at least PMIx v4.2.0
* The minimum required version for PRRTE is v3.0.0, but we recommend
  in the docs that users use >=v3.0.1 so that they get a full
  mpirun(1) man page

Finally, also show in the docs the versions of the embedded packages
(hwloc, libevent, PMIx, and PRRTE).  This required adding a little
Python in docs/conf.py to read VERSION files and extract version
numbers from tarball filenames.

Signed-off-by: Jeff Squyres <jeff@squyres.com>