Patch the prrte.spec file.
This is already fixed in prrte but for v5.0.x
and main we'll want this fix applied for any
rpm generation. This can safely be removed
once main and v5.0.x advance. On v5.0.x this will
be the next prrte release. For main, the next submodule
update is fine to remove this.

Signed-off-by: Austen Lauria <awlauria@us.ibm.com>