docs: Add info about building accelerators as DSOs
Add information about building the accelerator-related components as
DSOs, and an explanation of why pacakgers might want to do so.

We need these docs for v5.0.0; if some of these specifics change later
in the v5.0.x series (e.g., if we starting building these components
as DSOs by default), this documentation can be updated to match.

Signed-off-by: Jeff Squyres <jeff@squyres.com>